backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
marshmallow==3.21.3
python-dotenv==1.0.0
pytest==7.4.0
flask-cors==3.0.10
flasgger==0.9.7.1
flask-jwt-extended==4.7.1